+++ actix-http/examples/echo.rs
@@ -1,6 +1,6 @@
 use std::{env, io};
 
-use actix_http::{Error, HttpService, Request, Response};
+use actix_http::{http::StatusCode, Error, HttpService, Request, Response};
 use actix_server::Server;
 use bytes::BytesMut;
 use futures_util::StreamExt as _;
@@ -25,7 +25,7 @@ async fn main() -> io::Result<()> {
 
                     info!("request body: {:?}", body);
                     Ok::<_, Error>(
-                        Response::Ok()
+                        Response::build(StatusCode::OK)
                             .insert_header((
                                 "x-head",
                                 HeaderValue::from_static("dummy value!"),
