+++ actix-http/src/response.rs
@@ -29,18 +29,6 @@ pub struct Response<B> {
 }
 
 impl Response<Body> {
-    /// Create HTTP response builder with specific status.
-    #[inline]
-    pub fn build(status: StatusCode) -> ResponseBuilder {
-        ResponseBuilder::new(status)
-    }
-
-    /// Create HTTP response builder
-    #[inline]
-    pub fn build_from<T: Into<ResponseBuilder>>(source: T) -> ResponseBuilder {
-        source.into()
-    }
-
     /// Constructs a response
     #[inline]
     pub fn new(status: StatusCode) -> Response<Body> {
@@ -50,6 +38,41 @@ impl Response<Body> {
             error: None,
         }
     }
+
+    /// Create HTTP response builder with specific status.
+    #[inline]
+    pub fn build(status: StatusCode) -> ResponseBuilder {
+        ResponseBuilder::new(status)
+    }
+
+    // just a couple frequently used shortcuts
+    // this list should not grow larger than a few
+
+    /// Creates a new response with status 200 OK.
+    #[inline]
+    pub fn ok() -> Response<Body> {
+        Response::new(StatusCode::OK)
+    }
+
+    /// Creates a new response with status 400 Bad Request.
+    #[inline]
+    pub fn bad_request() -> Response<Body> {
+        Response::new(StatusCode::OK)
+    }
+
+    /// Creates a new response with status 404 Not Found.
+    #[inline]
+    pub fn not_found() -> Response<Body> {
+        Response::new(StatusCode::NOT_FOUND)
+    }
+
+    /// Creates a new response with status 500 Internal Server Error.
+    #[inline]
+    pub fn internal_server_error() -> Response<Body> {
+        Response::new(StatusCode::INTERNAL_SERVER_ERROR)
+    }
+
+    // end shortcuts
 
     /// Constructs an error response
     #[inline]
@@ -283,7 +306,7 @@ impl ResponseBuilder {
     /// # use actix_http::Response;
     /// use actix_http::http::header;
     ///
-    /// Response::Ok()
+    /// Response::build(StatusCode::OK)
     ///     .insert_header((header::CONTENT_TYPE, mime::APPLICATION_JSON))
     ///     .insert_header(("X-TEST", "value"))
     ///     .finish();
@@ -310,7 +333,7 @@ impl ResponseBuilder {
     /// # use actix_http::Response;
     /// use actix_http::http::header;
     ///
-    /// Response::Ok()
+    /// Response::build(StatusCode::OK)
     ///     .append_header((header::CONTENT_TYPE, mime::APPLICATION_JSON))
     ///     .append_header(("X-TEST", "value1"))
     ///     .append_header(("X-TEST", "value2"))
@@ -557,7 +580,7 @@ impl From<ResponseBuilder> for Response<Body> {
 
 impl From<&'static str> for Response<Body> {
     fn from(val: &'static str) -> Self {
-        Response::Ok()
+        Response::build(StatusCode::OK)
             .content_type(mime::TEXT_PLAIN_UTF_8)
             .body(val)
     }
@@ -565,7 +588,7 @@ impl From<&'static str> for Response<Body> {
 
 impl From<&'static [u8]> for Response<Body> {
     fn from(val: &'static [u8]) -> Self {
-        Response::Ok()
+        Response::build(StatusCode::OK)
             .content_type(mime::APPLICATION_OCTET_STREAM)
             .body(val)
     }
@@ -573,7 +596,7 @@ impl From<&'static [u8]> for Response<Body> {
 
 impl From<String> for Response<Body> {
     fn from(val: String) -> Self {
-        Response::Ok()
+        Response::build(StatusCode::OK)
             .content_type(mime::TEXT_PLAIN_UTF_8)
             .body(val)
     }
@@ -581,7 +604,7 @@ impl From<String> for Response<Body> {
 
 impl<'a> From<&'a String> for Response<Body> {
     fn from(val: &'a String) -> Self {
-        Response::Ok()
+        Response::build(StatusCode::OK)
             .content_type(mime::TEXT_PLAIN_UTF_8)
             .body(val)
     }
@@ -589,7 +612,7 @@ impl<'a> From<&'a String> for Response<Body> {
 
 impl From<Bytes> for Response<Body> {
     fn from(val: Bytes) -> Self {
-        Response::Ok()
+        Response::build(StatusCode::OK)
             .content_type(mime::APPLICATION_OCTET_STREAM)
             .body(val)
     }
@@ -597,7 +620,7 @@ impl From<Bytes> for Response<Body> {
 
 impl From<BytesMut> for Response<Body> {
     fn from(val: BytesMut) -> Self {
-        Response::Ok()
+        Response::build(StatusCode::OK)
             .content_type(mime::APPLICATION_OCTET_STREAM)
             .body(val)
     }
@@ -611,7 +634,7 @@ mod tests {
 
     #[test]
     fn test_debug() {
-        let resp = Response::Ok()
+        let resp = Response::build(StatusCode::OK)
             .append_header((COOKIE, HeaderValue::from_static("cookie1=value1; ")))
             .append_header((COOKIE, HeaderValue::from_static("cookie2=value2; ")))
             .finish();
@@ -621,7 +644,9 @@ mod tests {
 
     #[test]
     fn test_basic_builder() {
-        let resp = Response::Ok().insert_header(("X-TEST", "value")).finish();
+        let resp = Response::build(StatusCode::OK)
+            .insert_header(("X-TEST", "value"))
+            .finish();
         assert_eq!(resp.status(), StatusCode::OK);
     }
 
@@ -741,7 +766,7 @@ mod tests {
 
     #[test]
     fn response_builder_header_insert_kv() {
-        let mut res = Response::Ok();
+        let mut res = Response::build(StatusCode::OK);
         res.insert_header(("Content-Type", "application/octet-stream"));
         let res = res.finish();
 
@@ -753,7 +778,7 @@ mod tests {
 
     #[test]
     fn response_builder_header_insert_typed() {
-        let mut res = Response::Ok();
+        let mut res = Response::build(StatusCode::OK);
         res.insert_header((header::CONTENT_TYPE, mime::APPLICATION_OCTET_STREAM));
         let res = res.finish();
 
@@ -765,7 +790,7 @@ mod tests {
 
     #[test]
     fn response_builder_header_append_kv() {
-        let mut res = Response::Ok();
+        let mut res = Response::build(StatusCode::OK);
         res.append_header(("Content-Type", "application/octet-stream"));
         res.append_header(("Content-Type", "application/json"));
         let res = res.finish();
@@ -778,7 +803,7 @@ mod tests {
 
     #[test]
     fn response_builder_header_append_typed() {
-        let mut res = Response::Ok();
+        let mut res = Response::build(StatusCode::OK);
         res.append_header((header::CONTENT_TYPE, mime::APPLICATION_OCTET_STREAM));
         res.append_header((header::CONTENT_TYPE, mime::APPLICATION_JSON));
         let res = res.finish();
